feat(dom): refactoring location moved to corresponding folder/package
